requirements.txt
google-cloud-storage>=2.0.0
google-cloud-bigquery>=2.34.0
pandas>=1.3.0
requests>=2.25.0
python-dotenv>=0.19.0
pyspark>=3.1.0